+++ src/plugins/addServerTitlesSidebar/style.css
@@ -1,0 +1,94 @@
+/**
+ * @name addServerTitlesSidebar
+ * @description Extends the server list to show server titles next to the icons.
+ * @author KBO
+ * @version 1.0.0
+ */
+
+/* 
+ * By default, the server sidebar is only wide enough for the icons. 
+ * This rule expands it to make room for the server titles.
+ */
+nav[aria-label="Servers sidebar"] {
+    width: 240px !important;
+}
+
+/* 
+ * This targets server and folder items specifically, leaving other items 
+ * like the DMs button untouched. It overrides Discord's default centering
+ * and adds some padding to align the icons correctly in the wider sidebar.
+ * The :has() selector is used to only apply this to items that contain a server or folder.
+ */
+div[class*="listItem"]:has(div[class*="blobContainer"]),
+div[class*="listItem"]:has(div[class*="folderHeader"]) {
+    justify-content: flex-start !important;
+    padding-left: 17px;
+    box-sizing: border-box;
+}
+
+/*
+ * This rule widens the entire folder container, which includes the folder icon
+ * and its background. This ensures the background correctly wraps the title
+ * and the servers inside, in both collapsed and expanded states.
+ */
+div[class*="folderGroup"] {
+    width: 224px;
+}
+
+/* 
+ * These rules make the main containers for server icons and folder headers
+ * flexible, allowing the icon and the title to sit side-by-side.
+ */
+div[class*="blobContainer"],
+div[class*="folderHeader"] {
+    display: flex;
+    align-items: center;
+    width: 216px;
+}
+
+/* 
+ * This ensures the icon itself doesn't shrink when the title is added,
+ * and it uses the 'order' property to guarantee the icon always appears
+ * before the text, regardless of the HTML structure.
+ */
+div[class*="blobContainer"] > div[class*="wrapper"],
+div[class*="folderHeader"] > div[class*="folderButton"] {
+    flex-shrink: 0;
+    order: 1;
+}
+
+/* 
+ * This corrects a minor alignment issue with the folder button, ensuring
+ * it lines up perfectly with the server icons.
+ */
+div[class*="folderHeader"] > div[class*="folderButton"] {
+    margin-left: -4px;
+}
+
+/* 
+ * This defines the base styles for the injected server and folder titles,
+ * including font, color, and positioning. It also sets the 'order' property
+ * to ensure the title always appears after the icon.
+ */
+.server-name-text,
+.folder-name-text {
+    color: var(--header-primary);
+    font-family: var(--font-primary);
+    font-size: 15px;
+    font-weight: 500;
+    white-space: nowrap;
+    overflow: hidden;
+    text-overflow: ellipsis;
+    cursor: pointer;
+    margin-left: 17px;
+    width: 170px;
+    order: 2;
+}
+
+/* 
+ * This rule applies specific styling for folder titles to differentiate
+ * them from regular server titles, matching Discord's native look and feel.
+ */
+.folder-name-text {
+    margin-left: 13px !important;
+}
