--- src/plugins/serverTitlesSidebar/style.css
@@ -1,62 +1,0 @@
-/* Widen the server list panel */
-nav[aria-label="Servers sidebar"] {
-    width: 240px !important;
-}
-
-/*
- * Override Discord's centering and add padding, but ONLY for list items
- * that contain a server icon (blobContainer) or a folder.
-*/
-div[class*="listItem"]:has(div[class*="blobContainer"]),
-div[class*="listItem"]:has(div[class*="folderIconWrapper"]) {
-    justify-content: flex-start !important;
-    padding-left: 17px;
-    box-sizing: border-box;
-}
-
-/* Widen the entire folder container */
-div[class*="folderGroup"] {
-    width: 224px;
-}
-
-/* Make the containers flexible and wide to hold the icon and text */
-div[class*="blobContainer"],
-div[class*="folderHeader"] {
-    display: flex;
-    align-items: center;
-    /* This width is based on the sidebar width (240px) minus the item's padding (12px) and some extra for margin */
-    width: 216px;
-}
-
-/* Prevent the icon from shrinking and enforce order */
-div[class*="blobContainer"] > div[class*="wrapper"],
-div[class*="folderHeader"] > div[class*="folderButton"] {
-    flex-shrink: 0;
-    order: 1;
-}
-
-/* Specific alignment fix for the folder button */
-div[class*="folderHeader"] > div[class*="folderButton"] {
-    margin-left: -4px;
-}
-
-/* Style for the injected server name text */
-.server-name-text,
-.folder-name-text {
-    color: var(--header-primary);
-    font-family: var(--font-primary);
-    font-size: 15px;
-    font-weight: 500;
-    white-space: nowrap;
-    overflow: hidden;
-    text-overflow: ellipsis;
-    cursor: pointer;
-    margin-left: 17px;
-    width: 170px;
-    order: 2;
-}
-
-/* Style for folder headers */
-.folder-name-text {
-    margin-left: 13px !important;
-}
